+++ lib/subghz/protocols/bett.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_bett_feed(void* context, bool level, uint32_t durat
  * @param context Pointer to a SubGhzProtocolDecoderBETT instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_bett_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_bett_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderBETT.
