+++ lib/subghz/protocols/bin_raw.h
@@ -79,7 +79,7 @@ void subghz_protocol_decoder_bin_raw_feed(void* context, bool level, uint32_t du
  * @param context Pointer to a SubGhzProtocolDecoderBinRAW instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_bin_raw_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_bin_raw_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderBinRAW.
