+++ lib/subghz/protocols/came_atomo.h
@@ -81,7 +81,7 @@ void subghz_protocol_decoder_came_atomo_feed(void* context, bool level, uint32_t
  * @param context Pointer to a SubGhzProtocolDecoderCameAtomo instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_came_atomo_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_came_atomo_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderCameAtomo.
