+++ lib/subghz/protocols/clemsa.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_clemsa_feed(void* context, bool level, uint32_t dur
  * @param context Pointer to a SubGhzProtocolDecoderClemsa instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_clemsa_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_clemsa_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderClemsa.
