+++ lib/subghz/protocols/genie.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_genie_feed(void* context, bool level, uint32_t dura
  * @param context Pointer to a SubGhzProtocolDecoderGenie instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_genie_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_genie_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderGenie.
