+++ lib/subghz/protocols/holtek_ht12x.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_holtek_th12x_feed(void* context, bool level, uint32
  * @param context Pointer to a SubGhzProtocolDecoderHoltek_HT12X instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_holtek_th12x_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_holtek_th12x_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderHoltek_HT12X.
