+++ lib/subghz/protocols/ido.h
@@ -43,7 +43,7 @@ void subghz_protocol_decoder_ido_feed(void* context, bool level, uint32_t durati
  * @param context Pointer to a SubGhzProtocolDecoderIDo instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_ido_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_ido_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderIDo.
