+++ lib/subghz/protocols/marantec.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_marantec_feed(void* context, bool level, uint32_t d
  * @param context Pointer to a SubGhzProtocolDecoderMarantec instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_marantec_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_marantec_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderMarantec.
