+++ lib/subghz/protocols/nero_radio.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_nero_radio_feed(void* context, bool level, uint32_t
  * @param context Pointer to a SubGhzProtocolDecoderNeroRadio instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_nero_radio_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_nero_radio_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderNeroRadio.
