+++ lib/subghz/protocols/nice_flor_s.h
@@ -80,7 +80,7 @@ void subghz_protocol_decoder_nice_flor_s_feed(void* context, bool level, uint32_
  * @param context Pointer to a SubGhzProtocolDecoderNiceFlorS instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_nice_flor_s_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_nice_flor_s_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderNiceFlorS.
