+++ lib/subghz/protocols/phoenix_v2.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_phoenix_v2_feed(void* context, bool level, uint32_t
  * @param context Pointer to a SubGhzProtocolDecoderPhoenix_V2 instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_phoenix_v2_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_phoenix_v2_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderPhoenix_V2.
