+++ lib/subghz/protocols/princeton.h
@@ -82,7 +82,7 @@ void subghz_protocol_decoder_princeton_feed(void* context, bool level, uint32_t 
  * @param context Pointer to a SubGhzProtocolDecoderPrinceton instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_princeton_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_princeton_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderPrinceton.
