+++ lib/subghz/protocols/smc5326.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_smc5326_feed(void* context, bool level, uint32_t du
  * @param context Pointer to a SubGhzProtocolDecoderSMC5326 instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_smc5326_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_smc5326_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderSMC5326.
