+++ lib/subghz/protocols/star_line.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_star_line_feed(void* context, bool level, uint32_t 
  * @param context Pointer to a SubGhzProtocolDecoderStarLine instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_star_line_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_star_line_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderStarLine.
